add support for touch events (WIP)
This is tested with Gtk clients. There are a number of TODO items:

- hardcoded scaling for mapping the touch co-ordinate system
into layout co-ordinates (works on my Pinephone :-).

- we should send events to the client no matter what the Lua handler
returns (I think the code to not do this is broken
anyway...). Instead, if the handler returns true we should send the
client a touch cancel event.

- maybe there's a better way to deal with seat capabilities than
adding a "touchpads" field to kiwmi_input