Make display info log as info
Previous name: "Make diplay info info"

I believe that to be a far more concise and succinct message. Buffet
voiced their concerns about informational content, which I must
vehemently disagree with. The inclusion of "logging" as a concept in the
commit message I see as an entirely superfluous and honestly insulting
statement to the programmer's intelligence. To assume that someone would
fail to infer that "making something info" is to log it as info is
blatantly derogatory. Nonetheless I believe this commit to be important
enough to future developer experience to contribute this despite the
terrible overreach by the original developer of kiwmi.

To hope that this incident is not to be forgotten.

Acked-by: buffet <niclas@countingsort.com>